moved integration tests to their correct location
