+++ core/tests/account_service_tests.rs
@@ -1,3 +1,5 @@
+mod integration_test;
+
 #[cfg(test)]
 mod account_tests {
     use lockbook_core::client::Error;
@@ -10,7 +12,6 @@ mod account_tests {
         DefaultDbProvider, DefaultFileMetadataRepo, DefaultSyncService, ImportError,
     };
 
-    use crate::{random_username, test_config, test_db};
     use lockbook_core::model::account::Account;
     use lockbook_core::model::api::NewAccountError;
     use lockbook_core::repo::account_repo::AccountRepo;
@@ -19,6 +20,7 @@ mod account_tests {
     use lockbook_core::service::sync_service::SyncService;
     use rsa::{BigUint, RSAPrivateKey};
     use std::mem::discriminant;
+    use crate::integration_test::{test_db, random_username, test_config};
 
     #[test]
     fn create_account_successfully() {
