+++ core/tests/get_document_tests.rs
@@ -1,13 +1,19 @@
+mod integration_test;
+
+
+
 #[cfg(test)]
 mod get_document_tests {
-    use crate::{
-        aes_decrypt_str, aes_key, aes_str, generate_account, random_filename, rsa_key, sign,
+    use crate::integration_test::{
+        aes_decrypt_str, aes_key, aes_str, generate_account, random_filename, rsa_key, sign
     };
     use lockbook_core::client::{Client, ClientImpl, Error};
     use lockbook_core::model::api::*;
     use lockbook_core::model::crypto::*;
     use lockbook_core::service::crypto_service::{AesImpl, SymmetricCryptoService};
     use uuid::Uuid;
+
+   use crate::assert_matches;
 
     #[test]
     fn get_document() {
