+++ core/tests/get_public_key_tests.rs
@@ -1,11 +1,17 @@
+mod integration_test;
+
+
+
 #[cfg(test)]
 mod get_public_key_tests {
-    use crate::{aes_key, generate_account, random_username, rsa_key, sign};
+    use crate::integration_test::{aes_key, generate_account, random_username, rsa_key, sign};
     use lockbook_core::client::{Client, ClientImpl, Error};
     use lockbook_core::model::api::*;
     use lockbook_core::model::crypto::*;
     use lockbook_core::service::crypto_service::{AesImpl, SymmetricCryptoService};
     use uuid::Uuid;
+
+   use crate::assert_matches;
 
     #[test]
     fn get_public_key() {
