--- integration_tests/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "lockbook-test"
-version = "0.1.0"
-authors = ["Parth <parth@mehrotra.me>"]
-edition = "2018"
-
-[dependencies]
-lockbook_core = { path = "../core" }
-rsa = { version = "^0.2", features = ["serde1"] }
-serde_json = "1.0.44"
-uuid = { version = "0.8.1", features = ["v4"] }
